+++ python/wintask/__init__.py
@@ -1,4 +1,11 @@
 from .scheduler import TaskScheduler
-from .triggers import DailyTrigger, Weekday, WeeklyTrigger
+from .triggers import DailyTrigger, Month, MonthlyTrigger, Weekday, WeeklyTrigger
 
-__all__ = ["DailyTrigger", "TaskScheduler", "Weekday", "WeeklyTrigger"]
+__all__ = [
+	"DailyTrigger",
+	"Month",
+	"MonthlyTrigger",
+	"TaskScheduler",
+	"Weekday",
+	"WeeklyTrigger",
+]